Render envValueFrom before env in generic chart
It is possible to use other env variables in a variable but only already defined variables.
This is almost only useful if you need to template a normal env variable with the value from a variable from a secret. To make this work it is important that the envValueFrom variables which can load a variable from a secret are rendered before the env variables.