Set concurrencyPolicy of CronJobs
Allow to specify the concurrecnyPolicy for the backup and cleanup Jobs.
The default of Kubernetes is "Allow" but I think for this chart it makes sense
to set the default to "Forbid".
Bumped the major version because this is kind of an breaking change, even though
it probably breaks nothing.