feat: use maps instead of lists
Lists can not be merged which forces users to define all namespaces and
policies in a single file. This gets chaotic really fast. By replacing the
lists with maps we make it possible to merge several values files into
one.

This is a breaking change which requires every user to change their
values. An upgrade guide is included in the README.