fix!: require setting of service port name
The service port in `{{- $svcPort := .Values.service.port -}}` does not exist
anymore since 4.0.0.

Therefore, the `port.number` for ingresses was empty. As enabling the specification
of multiple ports requires the specification of ports on the ingress, this is now
required in the chart.

BREAKING CHANGE: You must now set ingress.hosts[*].paths[*].servicePortName.